Windows build fixes
General fixes for windows build: primarily, a windows version of the mmap code
