+++ tools/defn.tmpl
@@ -6,7 +6,7 @@ from libcpp.vector cimport vector
 
 {% for file_name, fns in fns %}
 # AUTO GENERATED FROM tools/defn.tmpl
-cdef extern from "{{file_name}}" namespace "ss::iter":
+cdef extern from {{repr(file_name)}} namespace "ss::iter":
     {% for fn in fns %}
     {{ fn }}
     {% endfor %}
@@ -14,7 +14,7 @@ cdef extern from "{{file_name}}" namespace "ss::iter":
 
 {% for iter in iters.values() %}
 # AUTO GENERATED FROM tools/defn.tmpl
-cdef extern from "{{iter.file_name}}" namespace "ss::iter":
+cdef extern from {{repr(iter.file_name)}} namespace "ss::iter":
     cdef cppclass {{iter.templated_name}}:
         {{iter.name}}({{iter.constructor_args}}) except +
         {{ iter.extra }}
